Use non-breaking space at end of line after since/depuis
Space characters tend to collapse at the end of cells in the first
column of the table. It happened notably in the latest French version
of the document, but this is not the first time. It happened before and
I have not identified which table or character style controls this behavior.

I have replaced the space characters with non-breaking spaces,
which have the same width but are reliably preserved in rendering.